+++ src/internal/ensure-iterable.mjs
@@ -4,9 +4,9 @@ export default function ensureIterable (i) {
     return emptyArr[Symbol.iterator]()
   } else if (!i[Symbol.iterator]) {
     if (typeof i.next === 'function') {
-      throw new Error('Iterators are not supported arguments to iter-tools. You must wrap them using the `iterable` method.')
+      throw new TypeError('Iterators are not supported arguments to iter-tools. You must wrap them using the `iterable` method.')
     }
-    throw new Error('The argument is not an iterable, or null')
+    throw new TypeError('The argument is not an iterable or null')
   }
   return i
 }