+++ src/slice.mjs
@@ -1,18 +1,43 @@
+import CircularBuffer from './internal/circular-buffer'
 import ensureIterable from './internal/ensure-iterable'
 
-function * slice (opts, iterable) {
-  let start, step, end
-  opts = typeof opts === 'number' ? { end: opts, start: 0 } : opts
+function bufferedSlice (iterable, start, end, step) {
+  const bufferSize = Math.abs(start)
 
-  step = typeof opts.step === 'undefined' ? 1 : opts.step
-  end = typeof opts.end === 'undefined'
-    ? (step > 0 ? Infinity : -Infinity) : opts.end
-  start = opts.start ? opts.start : 0
+  const buffer = new CircularBuffer(bufferSize)
+  for (let item of iterable) {
+    buffer.push(item)
+  }
+
+  let newEnd
+  if (isFinite(end) && end > 0) {
+    newEnd = end - (buffer.counter - bufferSize)
+    if (newEnd < 0) return []
+  } else {
+    newEnd = end
+  }
+  return simpleSlice(buffer, 0, newEnd, step)
+}
+
+function * simpleSlice (iterable, start, end, step) {
   let currentPos = 0
   let nextValidPos = start
+  const bufferSize = Math.abs(end)
+  let buffer
 
-  for (const item of ensureIterable(iterable)) {
-    if (currentPos >= end) {
+  if (end < 0) {
+    buffer = new CircularBuffer(bufferSize)
+  }
+
+  for (let item of iterable) {
+    if (buffer) {
+      item = buffer.push(item)
+      if (buffer.counter <= bufferSize) {
+        continue
+      }
+    }
+
+    if (currentPos >= end && end >= 0) {
       break
     }
 
@@ -21,6 +46,26 @@ function * slice (opts, iterable) {
       nextValidPos += step
     }
     currentPos++
+  }
+}
+
+function * slice (opts, iterable) {
+  let start, step, end
+  opts = typeof opts === 'number' ? { end: opts, start: 0 } : opts
+
+  step = typeof opts.step === 'undefined' ? 1 : opts.step
+  end = typeof opts.end === 'undefined' ? Infinity : opts.end
+  start = opts.start ? opts.start : 0
+  iterable = ensureIterable(iterable)
+
+  if (step <= 0) {
+    throw new TypeError('Cannot slice with step <= 0')
+  }
+
+  if (start >= 0) {
+    yield * simpleSlice(iterable, start, end, step)
+  } else {
+    yield * bufferedSlice(iterable, start, end, step)
   }
 }